OP-526: Switch flight side over to CRC-32 model

+++ flight/PiOS/inc/pios_crc.h
@@ -27,5 +27,5 @@
  * 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  */
 
-uint8_t PIOS_CRC_updateByte(uint8_t crc, const uint8_t data);
-uint8_t PIOS_CRC_updateCRC(uint8_t crc, const uint8_t* data, int32_t length);
+uint32_t PIOS_CRC_updateByte(uint32_t crc, const uint8_t data);
+uint32_t PIOS_CRC_updateCRC(uint32_t crc, const uint8_t* data, int32_t length);
